+++ examples/vkcs_compute_instance/main.tf
@@ -2,7 +2,6 @@ resource "vkcs_compute_instance" "compute" {
   name            = "compute-instance"
   flavor_id       = data.vkcs_compute_flavor.compute.id
   security_groups = ["default"]
-  image_id = data.vkcs_images_image.compute.id
   availability_zone = "GZ1"
 
   block_device {
@@ -20,7 +19,6 @@ resource "vkcs_compute_instance" "compute" {
     destination_type      = "volume"
     volume_type           = "ceph-ssd"
     volume_size           = 8
-    boot_index            = 1
     delete_on_termination = true
   }