+++ CHANGELOG.md
@@ -10,7 +10,7 @@ description: |-
 #### v0.2.2 (unreleased)
 - Remove "optional" property from ip attribute of DB instance
 - Fix error of concurrent file read/write in Image resource
-- Fix error of not expecting "to_delete" status when deleting DB backup
+- Fix errors related to unexpected new statuses for DB instance and DB backup resources
 
 #### v0.2.1
 - Support shards/instances/volumes resizing in DB cluster with shards