+++ CHANGELOG.md
@@ -7,9 +7,10 @@ description: |-
 
 # VKCS Provider's changelog
 
-#### v0.2.2 (unreleased)
+#### v0.2.2
 - Remove "optional" property from ip attribute of DB instance
 - Add validation and default value for block_device.boot_index of Compute instance resource
+- Structure the documentation by grouping it by service
 - Fix error of concurrent file read/write in Image resource
 - Fix error when creating DB instance replica with specified configuration_id
 - Fix errors related to unexpected new statuses for DB instance and DB backup resources